+++ examples/sync/src/net/request_id.rs
@@ -29,3 +29,24 @@ impl Generator {
         self.counter.fetch_add(1, Ordering::Relaxed)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::net::Generator;
+
+    #[test]
+    fn test_request_id_generation() {
+        let requester = Generator::new();
+        let id1 = requester.next();
+        let id2 = requester.next();
+        let id3 = requester.next();
+
+        // Request IDs should be monotonically increasing
+        assert!(id2 > id1);
+        assert!(id3 > id2);
+
+        // Should be consecutive since we're using a single Requester
+        assert_eq!(id2, id1 + 1);
+        assert_eq!(id3, id2 + 1);
+    }
+}
